requirements.txt
fastapi[all]>=0.111
uvicorn[standard]>=0.30
Pillow>=10.3
imageio
